feat: separate channel icon component
